fix assert type

+++ agent/engine/host_resource_manager_test.go
@@ -91,15 +91,15 @@ func TestHostResourceConsume(t *testing.T) {
 	taskResources := getTestTaskResourceMap(int64(512), int64(768), []*string{&taskPort1}, []*string{&taskPort2}, 1)
 
 	h.consume(testTaskArn, taskResources)
-	assert.Equal(t, *h.consumedResource["CPU"].IntegerValue, 512, "Incorrect cpu resource accounting during consume")
-	assert.Equal(t, *h.consumedResource["MEMORY"].IntegerValue, 768, "Incorrect memory resource accounting during consume")
+	assert.Equal(t, *h.consumedResource["CPU"].IntegerValue, int64(512), "Incorrect cpu resource accounting during consume")
+	assert.Equal(t, *h.consumedResource["MEMORY"].IntegerValue, int64(768), "Incorrect memory resource accounting during consume")
 	assert.Equal(t, h.consumedResource["PORTS"].StringSetValue[0], "22", "Incorrect port resource accounting during consume")
 	assert.Equal(t, h.consumedResource["PORTS"].StringSetValue[1], "23", "Incorrect port resource accounting during consume")
 	assert.Equal(t, len(h.consumedResource["PORTS"].StringSetValue), 2, "Incorrect port resource accounting during consume")
 	assert.Equal(t, *h.consumedResource["PORTS_UDP"].StringSetValue[0], "1000", "Incorrect udp port resource accounting during consume")
 	assert.Equal(t, *h.consumedResource["PORTS_UDP"].StringSetValue[1], "1001", "Incorrect udp port resource accounting during consume")
 	assert.Equal(t, len(h.consumedResource["PORTS_UDP"].StringSetValue), 2, "Incorrect port resource accounting during consume")
-	assert.Equal(t, *h.consumedResource["GPU"].IntegerValue, 3, "Incorrect gpu resource accounting during consume")
+	assert.Equal(t, *h.consumedResource["GPU"].IntegerValue, int64(3), "Incorrect gpu resource accounting during consume")
 }
 
 func TestHostResourceRelease(t *testing.T) {
@@ -116,13 +116,13 @@ func TestHostResourceRelease(t *testing.T) {
 	h.consume(testTaskArn, taskResources)
 	h.release(testTaskArn, taskResources)
 
-	assert.Equal(t, *h.consumedResource["CPU"].IntegerValue, 0, "Incorrect cpu resource accounting during release")
-	assert.Equal(t, *h.consumedResource["MEMORY"].IntegerValue, 0, "Incorrect memory resource accounting during release")
+	assert.Equal(t, *h.consumedResource["CPU"].IntegerValue, int64(0), "Incorrect cpu resource accounting during release")
+	assert.Equal(t, *h.consumedResource["MEMORY"].IntegerValue, int64(0), "Incorrect memory resource accounting during release")
 	assert.Equal(t, *h.consumedResource["PORTS"].StringSetValue[0], "22", "Incorrect port resource accounting during release")
 	assert.Equal(t, len(h.consumedResource["PORTS"].StringSetValue), 0, "Incorrect port resource accounting during release")
 	assert.Equal(t, *h.consumedResource["PORTS_UDP"].StringSetValue[0], "1000", "Incorrect udp port resource accounting during release")
 	assert.Equal(t, len(h.consumedResource["PORTS_UDP"].StringSetValue), 0, "Incorrect udp port resource accounting during release")
-	assert.Equal(t, *h.consumedResource["GPU"].IntegerValue, 4, "Incorrect gpu resource accounting during release")
+	assert.Equal(t, *h.consumedResource["GPU"].IntegerValue, int64(4), "Incorrect gpu resource accounting during release")
 }
 
 func TestConsumable(t *testing.T) {
